% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmzMLFunctions.R
\name{$.msdata_connection}
\alias{$.msdata_connection}
\title{S3 dollar sign notation for msdata_connection objects}
\usage{
\method{$}{msdata_connection}(msdata_obj, ms_level)
}
\arguments{
\item{msdata_obj}{An msdata_connection object containing files and grab_what}

\item{ms_level}{The requested MS level of the object}
}
\value{
An msdata_connection object with only a single MS level
}
\description{
S3 dollar sign notation for msdata_connection objects
}
