% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabMSdataCode.R
\name{checkFiles}
\alias{checkFiles}
\title{Check that filenames are acceptable.}
\usage{
checkFiles(files)
}
\arguments{
\item{files}{The vector of files provided to grabMSdata.}
}
\value{
NULL (invisibly). The goal of this function is its side effects, i.e.
  throwing errors and providing info when the files are not found.
}
\description{
Check that filenames are acceptable.
}
