% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabMSdataCode.R
\name{checkOutputQuality}
\alias{checkOutputQuality}
\title{Check that the output data is properly formatted.}
\usage{
checkOutputQuality(output_data, grab_what)
}
\arguments{
\item{output_data}{The collected data resulting from repeated calls to
`grabMzmlData()`, after being bound together.}

\item{grab_what}{The names of the data requested by the user.}
}
\value{
NULL (invisibly). The goal of this function is its side effects, i.e.
  throwing errors and providing info when the files are not found.
}
\description{
This function checks that data produced by repeated calls to the
`grabMzmlData()` and `grabMzxmlData()` functions is formatted properly before
it's provided to the user. It checks that all of the requested data has been
obtained and warns if data is found to be empty, misnamed, or has columns of
the wrong type.
}
