% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabData.R, R/grabMzmlCode.R
\name{grabMzmlData}
\alias{grabMzmlData}
\title{Read an mzML file into a data.frame}
\usage{
grabMzmlData(filename)

grabMzmlData(filename)
}
\arguments{
\item{filename}{The name of the mzML file to be read.}
}
\value{
A data.frame object with columns for retention time (rt) in minutes,
m/z (mz), and intensity (int).

A data.frame object with columns for retention time (rt) in minutes,
m/z (mz), and intensity (int).
}
\description{
Read an mzML file into a data.frame

Read an mzML file into a data.frame
}
\details{
This function reads an mzML file into R's working memory. mzML files
are fundamentally XML documents, which allows rapid access to the data by
parsing the XML. The R package `xml2::` is used for this purpose here.
Retention time information can be read directly, while *m/z* and intensity
information must be decoded from binary.

This function reads an mzML file into R's working memory. mzML files
are fundamentally XML documents, which allows rapid access to the data by
parsing the XML. The R package `xml2::` is used for this purpose here.
Retention time information can be read directly, while *m/z* and intensity
information must be decoded from binary.
}
\examples{
mzML_filename <- system.file("extdata", "190715_Poo_TruePooFK180310_Full2.mzML", package = "RaMS")
grabMzmlData(mzML_filename)
mzML_filename <- system.file("proteomics", "MS3TMT11.mzML", package = "msdata")
grabMzmlData(mzML_filename)
}
