% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabMzmlCode.R
\name{grabMzmlEIC}
\alias{grabMzmlEIC}
\title{Simple wrapper around grabMzmlData to extract a specific mass across multiple files}
\usage{
grabMzmlEIC(mass, ppm, filenames)
}
\arguments{
\item{mass}{The *m/z* of the compound of interest}

\item{ppm}{The instrument's parts-per-million accuracy}

\item{filenames}{A vector of complete paths to the mzML files to be read, i.e.
those produced by system.file() or list.files(full.names=TRUE)}
}
\value{
A data.frame object with columns for retention time (rt) in minutes,
*m/z*, intensity (int), and filename.
}
\description{
Simple wrapper around grabMzmlData to extract a specific mass across multiple files
}
\details{
This function reads an mzML file's *m/z*, retention time, and intensity
data into R using `grabMzmlData` and extracts a mass slice according to the user-
provided *m/z* and ppm. It is vectorized via `sapply` across all the files in
the `filenames` vector.
}
\examples{
msdata_1 <- system.file("proteomics", "MS3TMT10_01022016_32917-33481.mzML.gz", package = "msdata")
msdata_2 <- system.file("proteomics", "MS3TMT11.mzML", package = "msdata")
grabMzmlEIC(mass=444.22, ppm=50, filenames=c(msdata_1, msdata_2))
}
