% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabMzmlFunctions.R
\name{grabMzmlMetadata}
\alias{grabMzmlMetadata}
\title{Helper function to extract mzML file metadata}
\usage{
grabMzmlMetadata(xml_data)
}
\arguments{
\item{xml_data}{mzML data as parsed by xml2}
}
\value{
A list of values corresponding to various pieces of metadata
for each file
}
\description{
Helper function to extract mzML file metadata
}
