% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabMzmlCode.R
\name{grabMzmlMS2}
\alias{grabMzmlMS2}
\title{Read an mzML file's MS2 data into a data.frame}
\usage{
grabMzmlMS2(filename)
}
\arguments{
\item{filename}{The name of the mzML file to be read.}
}
\value{
A data.frame object with columns for retention time (rt) in minutes,
precursor mass (premz), fragment m/z (fragmz), and intensity (int).
}
\description{
Read an mzML file's MS2 data into a data.frame
}
\details{
This function reads an mzML file's MS2 data into R's working memory. mzML files
are fundamentally XML documents, which allows rapid access to the data by
parsing the XML. The R package `xml2::` is used for this purpose here.
Retention time and precursor mass information can be read directly, while m/z and intensity
information must be decoded from binary.
}
\examples{
sf <- "TMT_Erwinia_1uLSike_Top10HCD_isol2_45stepped_60min_01.mzML.gz"
mzML_MS2_filename <- system.file("proteomics", sf, package = "msdata")
grabMzmlMS2(mzML_MS2_filename)
}
