% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabData.R
\name{grabMzxmlData}
\alias{grabMzxmlData}
\title{Read an mzXML file into a data.frame}
\usage{
grabMzxmlData(filename)
}
\arguments{
\item{filename}{The name of the mzXML file to be read.}
}
\value{
A data.frame object with columns for retention time (rt) in minutes,
m/z (mz), and intensity (int).
}
\description{
Read an mzXML file into a data.frame
}
\details{
This function reads an mzXML file into R's working memory. mzXML files
are fundamentally XML documents, which allows rapid access to the data by
parsing the XML. The R package `xml2::` is used for this purpose here.
Retention time information can be read directly, while *m/z* and intensity
information must be decoded from binary. To read data from the mzML format
instead, check out grabMzmlData.
}
\examples{
mzXML_filename <- system.file("extdata", "190715_Poo_TruePooFK180310_Full2.mzXML", package = "RaMS")
grabMzxmlData(mzXML_filename)
}
