% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabMzxmlFunctions.R
\name{grabMzxmlEncodingData}
\alias{grabMzxmlEncodingData}
\title{Helper function to extract mzXML file metadata}
\usage{
grabMzxmlEncodingData(xml_data)
}
\arguments{
\item{xml_data}{mzXML data as parsed by xml2}
}
\value{
A list of values used by other parsing functions, currently
compression, precision, and endian encoding (endi_enc)
}
\description{
Helper function to extract mzXML file metadata
}
