% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabMzxmlFunctions.R
\name{grabMzxmlMetadata}
\alias{grabMzxmlMetadata}
\title{Helper function to extract mzXML file metadata}
\usage{
grabMzxmlMetadata(xml_data)
}
\arguments{
\item{xml_data}{mzXML data as parsed by xml2}
}
\value{
A list of values corresponding to various pieces of metadata
for each file
}
\description{
Helper function to extract mzXML file metadata
}
