% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabMzxmlFunctions.R
\name{grabMzxmlMS1}
\alias{grabMzxmlMS1}
\title{Extract the MS1 data from an mzXML nodeset}
\usage{
grabMzxmlMS1(xml_data, file_metadata, rtrange)
}
\arguments{
\item{xml_data}{An `xml2` nodeset, usually created by applying `read_xml` to
an mzXML file.}

\item{file_metadata}{Information about the file used to decode the binary
arrays containing m/z and intensity information.}
}
\value{
A `data.table` with columns for retention time (rt), m/z (mz),
and intensity (int).
}
\description{
Extract the MS1 data from an mzXML nodeset
}
