% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabMzxmlFunctions.R
\name{grabMzxmlMS2}
\alias{grabMzxmlMS2}
\title{Extract the MS2 data from an mzXML nodeset}
\usage{
grabMzxmlMS2(xml_data, file_metadata, rtrange)
}
\arguments{
\item{xml_data}{An `xml2` nodeset, usually created by applying `read_xml` to
an mzXML file.}

\item{file_metadata}{Information about the file used to decode the binary
arrays containing m/z and intensity information.}
}
\value{
A `data.table` with columns for retention time (rt),  precursor m/z (mz),
fragment m/z (fragmz), collision energy (voltage), and intensity (int).
}
\description{
Extract the MS2 data from an mzXML nodeset
}
