% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabMzxmlFunctions.R
\name{grabMzxmlSpectraRt}
\alias{grabMzxmlSpectraRt}
\title{Extract the retention time from the spectra of an mzXML nodeset}
\usage{
grabMzxmlSpectraRt(xml_nodes)
}
\arguments{
\item{xml_nodes}{An xml_nodeset object corresponding to the spectra collected
by the mass spectrometer, usually produced by applying `xml_find_all` to an
MS1 or MS2 nodeset.}
}
\value{
A numeric vector of retention times, one for each scan
}
\description{
Extract the retention time from the spectra of an mzXML nodeset
}
