% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabData.R
\name{grabSingleFileBPC}
\alias{grabSingleFileBPC}
\title{Grab the BPC or TIC for a given file}
\usage{
grabSingleFileBPC(filename, TIC = FALSE)
}
\arguments{
\item{filename}{The name of the mzML file to be read}

\item{TIC}{Should the BPC or TIC be read? If TIC=TRUE,
the total ion current (TIC)
for each retention time is read in. Otherwise, the BPC
(base peak chromatogram) is read for each retention time, corresponding
to the *maximum* intensity for each scan.}
}
\value{
A data.frame object with columns for retention time (rt) and
and intensity (int) corresponding to the
}
\description{
Grab the BPC or TIC for a given file
}
\examples{
mzML_filename <- "data/180205_Poo_TruePoo_Full2.mzML"
grabSingleFileBPC(mzML_filename)
grabSingleFileBPC(mzML_filename, TIC=TRUE)
}
