% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabData.R
\name{grabSingleFileData}
\alias{grabSingleFileData}
\title{Read an mzML file into a data.frame}
\usage{
grabSingleFileData(filename)
}
\arguments{
\item{filename}{The name of the mzML file to be read}
}
\value{
A data.frame object with columns for retention time (rt), m/z (mz),
and intensity (int)
}
\description{
Read an mzML file into a data.frame
}
\examples{
mzML_filename <- "data/180205_Poo_TruePoo_Full2.mzML"
grabSingleFileData(mzML_filename)
}
