% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabMzmlFunctions.R
\name{grabSpectraPremz}
\alias{grabSpectraPremz}
\title{Extract the precursor mass from the spectra of an mzML nodeset}
\usage{
grabSpectraPremz(xml_nodes)
}
\arguments{
\item{xml_nodes}{An xml_nodeset object corresponding to the spectra collected
by the mass spectrometer, usually produced by applying `xml_find_all` to an
MS1 or MS2 nodeset.}
}
\value{
A numeric vector of precursor masses, one for each scan
}
\description{
Extract the precursor mass from the spectra of an mzML nodeset
}
