% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabMzmlFunctions.R
\name{grabSpectraVoltage}
\alias{grabSpectraVoltage}
\title{Extract the collison energies from the spectra of an mzML nodeset}
\usage{
grabSpectraVoltage(xml_nodes)
}
\arguments{
\item{xml_nodes}{An xml_nodeset object corresponding to the spectra collected
by the mass spectrometer, usually produced by applying `xml_find_all` to an
MS1 or MS2 nodeset.}
}
\value{
A numeric vector of collision energies, one for each scan.
}
\description{
Although the collision energy is typically fixed per file, it's equally
fast (afaik) to just grab them all individually here. Also, I'm worried about
these rumors of "ramped" collision energies
}
