% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmzMLFunctions.R
\name{msdata_connection}
\alias{msdata_connection}
\title{S3 constructor for msdata_connection}
\usage{
msdata_connection(x)
}
\arguments{
\item{x}{This is a thing?}
}
\value{
Itself, with the class?
}
\description{
S3 constructor for msdata_connection
}
