% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmzMLFunctions.R
\name{node2dt}
\alias{node2dt}
\title{Convert node to data.table}
\usage{
node2dt(dubset_node, ms_level)
}
\arguments{
\item{dubset_node}{The "data subset" node with children rt, mz, etc.}

\item{ms_level}{The requested MS level to search for}
}
\value{
A data.table with columns depending on the MS level requested
}
\description{
Convert node to data.table
}
