% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabMzmlCode.R
\name{pmppm}
\alias{pmppm}
\title{Plus/Minus Parts Per Million}
\usage{
pmppm(mass, ppm = 4)
}
\arguments{
\item{mass}{The *m/z* of the molecule}

\item{ppm}{The instrumental error}
}
\value{
A length-two numeric vector
}
\description{
Plus/Minus Parts Per Million
}
\details{
Often, it's useful to convert a single *m/z* into an error range
based on the accuracy (in ppm) of the instrument that collected the data.
This function does so neatly and plays well with data.table's %between%
}
\examples{
pmppm(118.0865)
pmppm(118.0865, ppm=50)
pmppm(100)
pmppm(1000)
}
