% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmzMLFunctions.R
\name{print.msdata_connection}
\alias{print.msdata_connection}
\title{S3 print option for msdata_connection objects}
\usage{
\method{print}{msdata_connection}(x, ...)
}
\arguments{
\item{x}{An msdata_connection object containing files and grab_what}

\item{...}{Other arguments to be passed to print.default, I guess}
}
\value{
Messages, mostly
}
\description{
S3 print option for msdata_connection objects
}
