% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmzMLFunctions.R
\name{[.msdata_connection}
\alias{[.msdata_connection}
\title{S3 indexing for msdata_connection objects}
\usage{
\method{[}{msdata_connection}(msdata_obj, sub_func)
}
\arguments{
\item{msdata_obj}{An msdata_connection object containing files and grab_what}

\item{sub_func}{The function that will be parsed and used to subset the file}
}
\value{
A data.table with columns rt, mz, int, and filename
}
\description{
This is the step that actually performs the file opening and extraction!
}
