% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabMSdataCode.R
\name{grabMSdata}
\alias{grabMSdata}
\title{Grab mass-spectrometry data from file(s)}
\usage{
grabMSdata(
  files,
  grab_what = c("MS1", "MS2"),
  verbosity = "none",
  mz = NULL,
  ppm = NULL,
  rtrange = NULL,
  check_exists = TRUE
)
}
\arguments{
\item{files}{A character vector of filenames to read into R's memory. Both
absolute and relative paths are acceptable.}

\item{grab_what}{What data should be read from the file? Options include
"MS1" for data only from the first spectrometer, "MS2" for fragmentation
data, "BPC" for rapid access to the base peak chromatogram, and "TIC" for
rapid access to the total ion chromatogram. These options can be combined
(i.e. `grab_data=c("MS1", "MS2", "BPC")`) or this argument can be set to
"everything" to extract all of the above. Options "EIC" and "EIC_MS2" are
useful when working with files whose total size exceeds working memory -
they first extracts all relevant MS1 and MS2 data, then discard data
outside of the mass range(s) calculated from the provided mz and ppm.}

\item{verbosity}{Three levels of processing output to the R console: "very",
which provides information about each file as it's read in; "minimal", for
a progress bar but no individual file information; and "none" for no output
of any kind.}

\item{mz}{A vector of the mass-to-charge ratio for compounds of interest.
Only used when combined with `grab_what = "EIC"` (see above). Multiple
masses can be provided.}

\item{ppm}{A single number corresponding to the mass accuracy (in parts per
million) of the instrument on which the data was collected. Only used when
combined with `grab_what = "EIC"` (see above).}

\item{rtrange}{Only available when parsing mzML files. A vector of length 2
containing an upper and lower bound on retention times of interest.
Providing a range here can speed up load times (although not enormously, as
the entire file must still be read) and reduce the final object's size.}

\item{check_exists}{Boolean, default TRUE. Whether or not to check if a file
exists before attempting to load it. If TRUE and file cannot be found,
provides helpful warning message and stops function. Set to FALSE if
streaming files from internet repositories such as Metabolights.}
}
\value{
A list of `data.table`s, each named after the arguments requested in
  grab_what. $MS1 contains MS1 information, MS2 contains fragmentation info,
  etc. MS1 data has four columns: retention time (rt), mass-to-charge (mz),
  intensity (int), and filename. MS2 data has six: retention time (rt),
  precursor m/z (premz), fragment m/z (fragmz), fragment intensity (int),
  collision energy (voltage), and filename. Data requested that does not
  exist in the provided files (such as MS2 data requested from MS1-only
  files) will return an empty (length zero) data.table. The data.tables
  extracted from each of the individual files are collected into one large
  table using data.table's `rbindlist`.
}
\description{
The main `RaMS` function. This function accepts a list of the files that will
be read into R's working memory and returns a list of `data.table`s
containing the requested information. What information is requested is
determined by the `grab_what` argument, which can include MS1, MS2, BPC, TIC,
or metadata information. This function serves as a wrapper around both
`grabMzmlData` and `grabMzxmlData` and handles multiple files, but those two
have also been exposed to the user in case super-simple handling is desired.
}
\examples{
# Operations on a single file
sample_file <- system.file("extdata", "FK180310_Full1.mzML.gz",
                           package = "RaMS")
file_data <- grabMSdata(sample_file, grab_what="MS1")
# Extract MS1 data and a base peak chromatogram
file_data <- grabMSdata(sample_file, grab_what=c("MS1", "BPC"))
# Extract data from a retention time subset
file_data <- grabMSdata(sample_file, grab_what=c("MS1", "BPC"),
                        rtrange=c(5, 7))

# Get timing data
file_data <- grabMSdata(sample_file, verbosity="very")

# Extract data from multiple files simultaneously
sample_dir <- system.file("extdata", package = "RaMS")
sample_files <- list.files(sample_dir, pattern="Full1", full.names=TRUE)
multifile_data <- grabMSdata(sample_files, grab_what="MS1")

# Extract an EIC for glycine betaine from several files
sample_files <- list.files(sample_dir, pattern="mzML", full.names=TRUE)
bet_EIC <- grabMSdata(sample_files, grab_what="EIC", mz=118.0865, ppm=5)
# Or speed it up even more because you already know betaine's retention time
bet_EIC <- grabMSdata(sample_files, grab_what="EIC", rtrange=c(5, 7),
                      mz=118.0865, ppm=5)

# Extract EIC for multiple masses simultaneously, from multiple files
file_data <- grabMSdata(sample_files, grab_what="EIC", ppm=5,
                        mz=c(118.0865, 146.118104, 189.123918))

# Extract MS2 data and search for specific fragment masses
library(data.table)
MS2_file <- list.files(sample_dir, pattern="DDA.*mzML", full.names=TRUE)
file_data <- grabMSdata(MS2_file, grab_what=c("MS1", "MS2"))
fragment_mass <- 59.0739
frags <- file_data$MS2[fragmz\%between\%pmppm(fragment_mass, ppm=5)]

# Or search for neutral losses
neutral_loss <- 60.0205
frags <- file_data$MS2[(premz-fragmz)\%between\%pmppm(neutral_loss, ppm=5)]

# Just get the file's metadata
metadata <- grabMSdata(MS2_file, grab_what="metadata")
}
