% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabMzmlCode.R
\name{grabMzmlBPC}
\alias{grabMzmlBPC}
\title{Grab the BPC or TIC for a given mzML file}
\usage{
grabMzmlBPC(filename, TIC = FALSE)
}
\arguments{
\item{filename}{The name of the mzML file to be read.}

\item{TIC}{Should the BPC or TIC be read? If TIC=TRUE,
the total ion current (TIC)
for each retention time is read in. Otherwise, the BPC
(base peak chromatogram) is read for each retention time, corresponding
to the *maximum* intensity for each scan.}
}
\value{
A data.frame object with columns for retention time (rt) in minutes and
and intensity (int) corresponding to the TIC or BPC for a given file.
}
\description{
Grab the BPC or TIC for a given mzML file
}
\examples{
mzML_filename <- system.file("proteomics", "MS3TMT11.mzML", package = "msdata")
grabMzmlBPC(mzML_filename)
grabMzmlBPC(mzML_filename, TIC=TRUE)
}
