% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabMzmlFunctions.R
\name{grabMzmlData}
\alias{grabMzmlData}
\title{Get mass-spectrometry data from an mzML file}
\usage{
grabMzmlData(
  filename,
  grab_what,
  verbose = FALSE,
  mz = NULL,
  ppm = NULL,
  rtrange = NULL
)
}
\arguments{
\item{filename}{A single filename to read into R's memory. Both absolute and
relative paths are acceptable.}

\item{grab_what}{What data should be read from the file? Options include
"MS1" for data only from the first spectrometer, "MS2" for fragmentation
data, "BPC" for rapid access to the base peak chromatogram, and "TIC" for
rapid access to the total ion chromatogram. These options can be combined
(i.e. `grab_data=c("MS1", "MS2", "BPC")`) or this argument can be set to
"everything" to extract all of the above. Option "EIC" is useful when
working with files whose total size exceeds working memory - it first
extracts all relevant MS1 and MS2 data, then discards data outside of the
mass range(s) calculated from the provided mz and ppm.}

\item{verbose}{Boolean. If TRUE, R will print information about timing to the
console as the file is read in.}

\item{mz}{A vector of the mass-to-charge ratio for compounds of interest.
Only used when combined with `grab_what = "EIC"` (see above). Multiple
masses can be provided.}

\item{ppm}{A single number corresponding to the mass accuracy (in parts per
million) of the instrument on which the data was collected. Only used when
combined with `grab_what = "EIC"` (see above).}

\item{rtrange}{A vector of length 2 containing an upper and lower bound on
retention times of interest. Providing a range here can speed up load times
(although not enormously, as the entire file must still be read) and reduce
the final object's size.}
}
\value{
A list of `data.table`s, each named after the arguments requested in
  grab_what. $MS1 contains MS1 information, $MS2 contains fragmentation info,
  etc. MS1 data has three columns: retention time (rt), mass-to-charge (mz),
  and intensity (int). MS2 data has five: retention time (rt), precursor m/z
  (premz), fragment m/z (fragmz), fragment intensity (int), and collision
  energy (voltage). Data requested that does not exist in the provided files
  (such as MS2 data requested from MS1-only files) will return an empty
  (length zero) data.table.
}
\description{
This function handles the mzML side of things, reading in files that are
written in the mzML format. Much of the code is similar to the mzXML format,
but the xpath handles are different and the mz/int array is encoded as two
separate entries rather than simultaneously.
}
