% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabMzmlCode.R
\name{grabMzmlMetadata}
\alias{grabMzmlMetadata}
\title{Helper function to extract mzML file metadata}
\usage{
grabMzmlMetadata(xml_data)
}
\arguments{
\item{xml_data}{mzML data as parsed by xml2}
}
\value{
A list of values used by other parsing functions, currently
compression, mz_precision, int_precision
}
\description{
Helper function to extract mzML file metadata
}
