% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabData.R
\name{grabMzxmlBPC}
\alias{grabMzxmlBPC}
\title{Grab the BPC or TIC for a given mzXML file}
\usage{
grabMzxmlBPC(filename, TIC = FALSE)
}
\arguments{
\item{filename}{The name of the mzXML file to be read.}

\item{TIC}{Should the BPC or TIC be read? If TIC=TRUE,
the total ion current (TIC)
for each retention time is read in. Otherwise, the BPC
(base peak chromatogram) is read for each retention time, corresponding
to the *maximum* intensity for each scan.}
}
\value{
A data.frame object with columns for retention time (rt) in minutes and
and intensity (int) corresponding to the TIC or BPC for a given file.
}
\description{
Grab the BPC or TIC for a given mzXML file
}
\examples{
mzXML_filename <- system.file("extdata", "190715_Poo_TruePooFK180310_Full2.mzXML", package = "RaMS")
grabMzxmlBPC(mzXML_filename)
grabMzxmlBPC(mzXML_filename, TIC=TRUE)
}
