% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabData.R, R/grabMzxmlCode.R
\name{grabMzxmlMetadata}
\alias{grabMzxmlMetadata}
\title{Helper function to extract mzXML file metadata}
\usage{
grabMzxmlMetadata(xml_data)

grabMzxmlMetadata(xml_data)
}
\arguments{
\item{xml_data}{mzXML data parsed by xml2}
}
\value{
A list of values used by other parsing functions, such as
compression type and precision

A list of values used by other parsing functions, such as
compression type and precision
}
\description{
Helper function to extract mzXML file metadata

Helper function to extract mzXML file metadata
}
