% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraHelperFunctions.R
\name{pmppm}
\alias{pmppm}
\title{Plus/minus parts per million}
\usage{
pmppm(mass, ppm = 4)
}
\arguments{
\item{mass}{A length-1 numeric representing the mass of
interest for which a mass range is desired.}

\item{ppm}{The parts-per-million accuracy of the mass spectrometer on which
the data was collected.}
}
\value{
A length-2 numeric representing the mass range requested
}
\description{
It shouldn't be hard to translate a point mass into a mass window bounded by
spectrometer accuracy.
}
\examples{
pmppm(100, 5)
pmppm(1000000, 5)
pmppm(118.0865, 2.5)
pmppm(892.535313, 10)
}
